% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genlasso-wrapper.R
\name{genlasso_wrapper}
\alias{genlasso_wrapper}
\title{Wrapper for \code{genlassoRcpp}}
\usage{
genlasso_wrapper(y, W, p, eta1, eta2, a, rho, max_iter, eps, truncate)
}
\description{
See for details \code{\link{genlassoRcpp}}
}
\seealso{
\code{\link{genlassoRcpp}}
}
