% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflsa.R
\name{get_estimate_lambda1}
\alias{get_estimate_lambda1}
\title{Get Estimates for Beta Coefficients for a Given Lambda1 Value}
\usage{
get_estimate_lambda1(fit, lambda1)
}
\arguments{
\item{fit}{The result object obtained from wflsa function}

\item{lambda1}{The lambda1 value for which beta coefficients are needed}
}
\value{
A list containing:
  \item{betas}{Matrix of beta coefficients for each lambda2 value. Each row corresponds to a single lambda2 value}
  \item{lambda2}{Vector of lambda2 values considered in the fit.}
}
\description{
This function takes the resulting fit from the wflsa function and returns the
estimates for the beta coefficients for all lambda2 values that were considered
before.
}
\examples{
# Example usage:
# fit <- wflsa(data, lambda1_values, lambda2_values)
# estimate <- get_estimate_lambda1(fit, lambda1_value)
# Now you can access estimated beta coefficients for different lambda2 values
}
