% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-unique-values.R
\name{get_unique_values}
\alias{get_unique_values}
\title{Get unique absolute values of a vector after rounding}
\usage{
get_unique_values(vector, digits = 10)
}
\arguments{
\item{vector}{A numeric vector.}

\item{digits}{Number of digits to round to (default is 10).}
}
\value{
A numeric vector containing the sorted unique absolute values.
}
\description{
This function takes a numeric vector, calculates its absolute values,
rounds them to a specified number of digits, finds the unique values,
sorts them, and returns the sorted unique absolute values.
}
\examples{
vector <- c(1.23456789, -1.23456788, 1.23456787, -1.23456786, 1.23456785, -1.23456784)
sorted_unique_abs_vals <- get_sorted_unique_abs_values(vector)
print(sorted_unique_abs_vals)
}
