% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfla.R
\name{print.wflsa.fit}
\alias{print.wflsa.fit}
\title{Print Function for the Weighted Fused LASSO Signal Approximator (wFLSA) fit object}
\usage{
\method{print}{wflsa.fit}(fit, ...)
}
\arguments{
\item{fit}{An object of class wflsa.fit.}

\item{...}{Additional arguments to be passed to the print function.}
}
\description{
This function prints information about the fitted wFLSA object, including the 
number of variables, the number of lambda pairs, and the estimated beta coefficients.
}
