% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflsa-package.R
\docType{package}
\name{wflsa-package}
\alias{wflsa-package}
\title{wflsa: Weighted Fused LASSO Signal Approximator ('wFLSA')}
\description{
A package for computing the Weighted Fused LASSO Signal Approximator (wFLSA).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bips-hb/wflsa}
  \item Report bugs at \url{https://github.com/bips-hb/wflsa/issues}
}

}
\author{
\strong{Maintainer}: Louis Dijkstra \email{dijkstra@leibniz-bips.de}

}
\keyword{internal}
