% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflsa.R
\name{wflsa}
\alias{wflsa}
\title{The Weighted Fused Lasso Signal Approximator (wFLSA) Algorithm}
\usage{
wflsa(
  y,
  W,
  lambda1 = c(0.1),
  lambda2 = c(0.1),
  rho = 1,
  max_iter = 1e+05,
  eps = 1e-10,
  truncate = 1e-04,
  offset = TRUE
)
}
\arguments{
\item{y}{Vector of length \eqn{p} representing the response variable (assumed to be centered).}

\item{W}{Weight matrix of dimensions \eqn{p \times p}.}

\item{lambda1}{Vector of positive regularization parameters for \eqn{L_1} penalty.}

\item{lambda2}{Vector of positive regularization parameters for smoothness penalty.}

\item{rho}{ADMM's parameter (Default: \code{1}).}

\item{max_iter}{Maximum number of iterations (Default: \code{1e5}).}

\item{eps}{Stopping criterion. If differences are smaller than \code{eps}, 
the algorithm halts (Default: \code{1e-10}).}

\item{truncate}{Values below \code{truncate} are set to \code{0} (Default: \code{1e-4}).}

\item{offset}{Logical indicating whether to include an intercept term (Default: \code{TRUE}).}
}
\value{
A list containing:
\itemize{
 \item \code{betas}: Estimated vector \eqn{\hat{\beta}} from the Weighted Fused LASSO.
 \item \code{tuning_parameters}: Data frame with tuning parameters. The column \code{df} 
       contains the number of non-zero coefficients for the different lambda-values
 \item all input variables.
}
}
\description{
Solves the weighted Fused LASSO Signal Approximator optimization problem 
using an ADMM-based approach. The problem is formulated as follows: 
\deqn{
\hat{\beta} = \argmin \frac{1}{2} || y - \beta ||_2^2 + \lambda_1 ||\beta ||_1 + \lambda_2 \sum_{i < j} w_{ij} | \beta_i - \beta_j |  
}
where:
\itemize{
 \item \eqn{y} is the response with mean \eqn{0}.
 \item \eqn{\beta} is the vector of coefficients to be estimated.
 \item \eqn{|| \cdot ||_1} and \eqn{|| \cdot ||_2} are the \eqn{L_1}- and \eqn{L_2}-norms, respectively.
 \item \eqn{\lambda_1 > 0} is the regularization parameter controlling the strength of the sparsity penalty. 
 \item \eqn{\lambda_2 > 0} is the regularization parameter controlling the smoothness. 
 \item \eqn{w_{ij} \in [0,1]} is the weight between the \eqn{i}-th and \eqn{j}-th coefficient.
}
}
\note{
\strong{Important Note:} The algorithm assumes \eqn{y} to be centered, i.e., its mean is 0.
}
\examples{
# Example usage of the wflsa function
y <- c(1, 2, 3)
W <- matrix(c(1, 0, 0, 0, 1, 0, 0, 0, 1), ncol = 3)
lambda1 <- c(0.1, 0.2)
lambda2 <- c(0.1, 0.2)
result <- wflsa(y, W, lambda1, lambda2)

}
\seealso{
\code{\link{genlassoRcpp}}
}
