% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflsa-package.R
\docType{package}
\name{wflsa-package}
\alias{wflsa-package}
\title{wflsa: Weighted Fused LASSO Signal Approximator ('wFLSA')}
\description{
A package for computing the Weighted Fused LASSO Signal Approximator (wFLSA).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bips-hb/wflsa}
  \item \url{https://bips-hb.github.io/wflsa/}
  \item Report bugs at \url{https://github.com/bips-hb/wflsa/issues}
}

}
\author{
\strong{Maintainer}: Ronja Foraita \email{foraita@leibniz-bips.de} (\href{https://orcid.org/0000-0003-2216-6653}{ORCID})

Authors:
\itemize{
  \item Louis Dijkstra (\href{https://orcid.org/0000-0003-1476-6202}{ORCID})
}

Other contributors:
\itemize{
  \item Leibniz Institute for Prevention Research and Epidemiology - BIPS (02c22vc57) [copyright holder, funder]
  \item Deutsche Forschungsgemeinschaft (DFG) [funder]
}

}
\keyword{internal}
